requirements.txt
Flask==3.1.3
gunicorn==21.2.0
google-cloud-bigquery==3.41.0
google-genai==1.3.0
pydantic==2.13.2
pytest==8.3.3
pytest-cov==6.0.0
pytest-mock==3.14.0
Flask-Cors==5.0.0
google-cloud-storage==2.18.2
